Redid the layout of the user-server
- The user-server used to be built on the code that Express
  auto-generates for a traditional web app. However, this was
  overkill for our project, which is a relatively simple REST
  API.
- Removed all the clutter/unnecessary files and directories.
- Moved/renamed some files and directories to be more logically
  organized. Namely, introduced lib/, which contains all the
  libraries we write for the user-server (e.g. dal.js, aps.js).
