+++ user-server/delete-database.sql
@@ -1,0 +1,11 @@
+# delete-database.sql
+# Author: Michael Friedman
+#
+# Drops all tables in the database. A quick way to reset the database
+# from scratch.
+
+drop table following;
+drop table followers;
+drop table profile_info;
+drop table posts;
+drop table photos;
