+++ user-server/package.json
@@ -1,26 +1,23 @@
 {
-  "name": "ibob2",
-  "version": "0.0.0",
-  "private": true,
+  "name": "user-server",
+  "version": "1.0.0",
+  "description": "",
+  "main": "server.js",
   "scripts": {
-    "start": "node ./bin/www",
-    "devstart": "nodemon ./bin/www",
+    "devstart": "nodemon node server.js",
+    "start": "node server.js",
     "test": "./node_modules/.bin/mocha --exit"
   },
+  "author": "",
+  "license": "ISC",
   "dependencies": {
     "axios": "^0.17.1",
-    "body-parser": "~1.18.2",
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.6.9",
-    "express": "~4.15.5",
-    "jade": "~1.11.0",
+    "express": "^4.16.2",
     "jsontokens": "^0.7.6",
-    "morgan": "~1.9.0",
-    "mysql": "^2.15.0",
-    "serve-favicon": "~2.4.5"
+    "mysql": "^2.15.0"
   },
   "devDependencies": {
     "mocha": "^4.0.1",
-    "nodemon": "^1.12.1"
+    "nodemon": "^1.14.3"
   }
 }
